Consolidate server-side modules
