+++ deployment/compute-cloud-sql/src/main/java/io/spine/examples/todolist/server/ComputeCloudSqlServer.java
@@ -25,7 +25,6 @@ import com.google.common.flogger.FluentLogger;
 import com.zaxxer.hikari.HikariConfig;
 import com.zaxxer.hikari.HikariDataSource;
 import io.spine.base.Environment;
-import io.spine.examples.todolist.TodoListContext;
 import io.spine.server.BoundedContext;
 import io.spine.server.ServerEnvironment;
 import io.spine.server.storage.StorageFactory;
