+++ deployment/local-firebase/src/main/java/io/spine/examples/todolist/server/LocalFirebaseServer.java
@@ -22,7 +22,6 @@ package io.spine.examples.todolist.server;
 
 import com.google.cloud.firestore.Firestore;
 import com.google.common.collect.ImmutableSet;
-import io.spine.examples.todolist.TodoListContext;
 import io.spine.examples.todolist.view.TaskView;
 import io.spine.server.BoundedContext;
 import io.spine.server.firebase.FirebaseSubscriptionMirror;
