+++ deployment/local-inmem/src/main/java/io/spine/examples/todolist/server/LocalInMemoryServer.java
@@ -23,8 +23,8 @@ package io.spine.examples.todolist.server;
 import java.io.IOException;
 
 import static io.spine.client.ConnectionConstants.DEFAULT_CLIENT_SERVICE_PORT;
-import static io.spine.examples.todolist.TodoListContext.create;
 import static io.spine.examples.todolist.server.Server.newServer;
+import static io.spine.examples.todolist.server.TodoListContext.create;
 
 /**
  * A local {@link Server} using
