+++ integration-tests/src/test/java/io/spine/test/AbstractIntegrationTest.java
@@ -23,13 +23,13 @@ package io.spine.test;
 import com.google.common.collect.ImmutableList;
 import com.zaxxer.hikari.HikariConfig;
 import com.zaxxer.hikari.HikariDataSource;
-import io.spine.examples.todolist.TodoListContext;
 import io.spine.examples.todolist.client.TodoClient;
 import io.spine.examples.todolist.client.builder.CommandBuilder;
 import io.spine.examples.todolist.command.CreateBasicLabel;
 import io.spine.examples.todolist.command.CreateBasicTask;
 import io.spine.examples.todolist.command.CreateDraft;
 import io.spine.examples.todolist.server.Server;
+import io.spine.examples.todolist.server.TodoListContext;
 import io.spine.server.BoundedContext;
 import io.spine.server.ServerEnvironment;
 import io.spine.server.storage.StorageFactory;
