+++ server/src/main/java/io/spine/examples/todolist/server/label/LabelAggregateRepository.java
