+++ server/src/main/java/io/spine/examples/todolist/server/label/LabelViewRepository.java
