+++ server/src/main/java/io/spine/examples/todolist/server/label/TaskLabelsPartRejections.java
