+++ server/src/main/java/io/spine/examples/todolist/server/label/package-info.java
