+++ server/src/main/java/io/spine/examples/todolist/server/task/TaskViewRepository.java
