+++ server/src/main/java/io/spine/examples/todolist/server/task/WizardCommands.java
