+++ server/src/test/java/io/spine/examples/todolist/server/label/LabelAggregateRejectionsTest.java
