+++ server/src/test/java/io/spine/examples/todolist/server/label/LabelAggregateTest.java
