+++ server/src/test/java/io/spine/examples/todolist/server/task/CompleteTaskTest.java
