+++ server/src/test/java/io/spine/examples/todolist/server/task/ReopenTaskCommandTest.java
